[ELY-1736] Clean up the 'wildfly-elytron-credential' module as it currently contains both public and private API [ELY-1736] Move 'org.wildfly.security.keystore' to its own module
