+++ credential/base/src/main/java/org/wildfly/security/credential/_private/ElytronMessages.java
@@ -91,101 +91,11 @@ public interface ElytronMessages extends BasicLogger {
     @Message(id = 1063, value = "Private key is null")
     IllegalArgumentException privateKeyIsNull();
     
-    @Message(id = 2001, value = "Invalid key store entry password for alias \"%s\"")
-    UnrecoverableKeyException invalidKeyStoreEntryPassword(String alias);
-    
-    @Message(id = 2002, value = "Invalid key store entry type for alias \"%s\" (expected %s, got %s)")
-    KeyStoreException invalidKeyStoreEntryType(String alias, Class<?> expectedClass, Class<?> actualClass);
-    
-    @Message(id = 2003, value = "Key store key for alias \"%s\" cannot be protected")
-    KeyStoreException keyCannotBeProtected(String alias);
-    
-    @Message(id = 2004, value = "Key store failed to translate password for alias \"%s\"")
-    IOException keyStoreFailedToTranslate(String alias, @Cause Throwable cause);
-    
-    @Message(id = 2005, value = "Key store failed to identify a suitable algorithm for alias \"%s\"")
-    NoSuchAlgorithmException noAlgorithmForPassword(String alias);
-    
-    @Message(id = 2006, value = "Unexpected whitespace in password file")
-    IOException unexpectedWhitespaceInPasswordFile();
-    
-    @Message(id = 2007, value = "Unexpected end of file")
-    EOFException unexpectedEof();
-    
-    @Message(id = 2008, value = "A reversible load is not possible until the KeyStore has first been initialized")
-    IllegalStateException reversibleLoadNotPossible();
-    
-    @Message(id = 2009, value = "Unable to create a new KeyStore instance")
-    IOException unableToCreateKeyStore(@Cause Exception cause);
-    
-    @Message(id = 2012, value = "An empty alias filter was supplied")
-    IllegalArgumentException emptyFilter();
-    
-    @Message(id = 2013, value = "Filter is missing '+' or '-' at offset %d")
-    IllegalArgumentException missingPlusMinusAt(long position);
-    
-    @Message(id = 2014, value = "Invalid first word '%s', must be one of ALL/NONE")
-    IllegalArgumentException invalidFirstWord(String firstWord);
-    
-    @Message(id = 2015, value = "Failed to obtain DirContext")
-    IllegalStateException failedToObtainDirContext(@Cause Throwable cause);
-    
-    @Message(id = 2016, value = "Failed to return DirContext")
-    IllegalStateException failedToReturnDirContext(@Cause Throwable cause);
-    
-    @Message(id = 2017, value = "LdapKeyStore failed to obtain alias [%s]")
-    IllegalStateException ldapKeyStoreFailedToObtainAlias(String alias, @Cause Throwable cause);
-    
-    @Message(id = 2018, value = "LdapKeyStore failed to obtain certificate [%s]")
-    IllegalStateException ldapKeyStoreFailedToObtainCertificate(String alias, @Cause Throwable cause);
-    
-    @Message(id = 2019, value = "LdapKeyStore failed to obtain certificate chain [%s]")
-    IllegalStateException ldapKeyStoreFailedToObtainCertificateChain(String alias, @Cause Throwable cause);
-    
-    @Message(id = 2020, value = "LdapKeyStore failed to recover key of alias [%s]")
-    IllegalStateException ldapKeyStoreFailedToObtainKey(String alias, @Cause Throwable cause);
-    
-    @Message(id = 2021, value = "LdapKeyStore failed to obtain alias by certificate")
-    IllegalStateException ldapKeyStoreFailedToObtainAliasByCertificate(@Cause Throwable cause);
-    
-    @Message(id = 2022, value = "LdapKeyStore failed to recover key of alias [%s]")
-    UnrecoverableKeyException ldapKeyStoreFailedToRecoverKey(String alias, @Cause Throwable cause);
-    
-    @Message(id = 2023, value = "LdapKeyStore failed to obtain creation date of alias [%s]")
-    IllegalStateException ldapKeyStoreFailedToObtainCreationDate(String alias, @Cause Throwable cause);
-    
-    @Message(id = 2024, value = "Alias [%s] does not exist in LdapKeyStore and not configured for creation")
-    KeyStoreException creationNotConfigured(String alias);
-    
-    @Message(id = 2025, value = "LdapKeyStore failed store alias [%s]")
-    KeyStoreException ldapKeyStoreFailedToStore(String alias, @Cause Throwable cause);
-    
-    @Message(id = 2026, value = "LdapKeyStore failed to serialize certificate of alias [%s]")
-    KeyStoreException ldapKeyStoreFailedToSerializeCertificate(String alias, @Cause Throwable cause);
-    
-    @Message(id = 2027, value = "LdapKeyStore failed to protect (pack into keystore) key of alias [%s]")
-    KeyStoreException ldapKeyStoreFailedToSerializeKey(String alias, @Cause Throwable cause);
-    
-    @Message(id = 2028, value = "LdapKeyStore failed to delete alias [%s]")
-    KeyStoreException ldapKeyStoreFailedToDelete(String alias, @Cause Throwable cause);
-    
-    @Message(id = 2029, value = "LdapKeyStore failed to delete alias [%s] - alias not found")
-    KeyStoreException ldapKeyStoreFailedToDeleteNonExisting(String alias);
-    
-    @Message(id = 2030, value = "LdapKeyStore failed to test alias [%s] existence")
-    IllegalStateException ldapKeyStoreFailedToTestAliasExistence(String alias, @Cause Throwable cause);
-    
-    @Message(id = 2031, value = "LdapKeyStore failed to iterate aliases")
-    IllegalStateException ldapKeyStoreFailedToIterateAliases(@Cause Throwable cause);
-    
     @Message(id = 2032, value = "keySpec must be SecretKeySpect, given: [%s]")
     InvalidKeySpecException keySpecMustBeSecretKeySpec(String type);
     
     @Message(id = 2033, value = "key must implement SecretKeySpec and keySpec must be SecretKeySpec, given key, keySpec: [%s]")
     InvalidKeySpecException keyMustImplementSecretKeySpecAndKeySpecMustBeSecretKeySpec(String keyAndKeySpec);
-
-    @Message(id = 2035, value = "KeyStore type could not be detected")
-    KeyStoreException keyStoreTypeNotDetected();
     
     // Also used in SASL / OTP
     @Message(id = 5151, value = "Invalid OTP algorithm \"%s\"")
