+++ keystore/src/main/java/org/wildfly/security/keystore/FilteringKeyStoreSpi.java
@@ -16,9 +16,6 @@
  * limitations under the License.
  */
 package org.wildfly.security.keystore;
-
-import org.wildfly.security.credential._private.ElytronMessages;
-
 import java.io.IOException;
 import java.io.InputStream;
 import java.security.Key;
